gps: fix JNI related AndroidRuntime crash
Since gingerbread, every JNI thread has to be "allocated" through the
GpsCallbacks::create_thread_cb to give the HAL a chance to store the
JNIEnv ptr in the threads local storage which is in turn fetched
again on any GpsCallbacks issue. Unfortunately the retrieval does
not check for NULL.

The librpc based threads issueing these callbacks via update_gps_*() are
not known to the VM and thus crashed AndroidRuntime.

We now queue up the rpc events and signal the status thread to send
them off via the callbacks. (There's some room for improvement as
we could save a memcpy by having gps-rpc allocate the update events
so we don't have to copy them for the queue again.)

Otherwise some code cleanups, less verbosity, some consolidation,
a fix for the scan interval which changed from seconds to milliseconds
(thx bzo for steering me into the right place), a lower maximum scan
interval as 8 seconds are definitvely not working out and it looks like
topa and rhod like to sleep while gps is still enabled.